requirements.txt
robotframework
robotframework-appiumlibrary
Appium-Python-Client
allure-robotframework
pytest
playwright
pytest-playwright
python-dotenv
PyYAML
requests
